src/backend/requirement.txt
#this creates the virtual environment.
python -m venv env 
env\Scripts\activate
flask_cors